moved few central ui widgets into widgets.py without changing functionality
