+++ .gitignore
@@ -167,3 +167,6 @@ cython_debug/
 # user specific config files
 *.ini
 *.json
+
+# data files
+*.dat
