--- 20251016_152757_Example device.dat
@@ -1,3 +1,0 @@
-YYYY.MM.DD hh:mm:ss,Random value (0-100)
-2025.10.16 15:27:57,61.29
-2025.10.16 15:27:58,33.81
