--- 20251016_152757_test_RHTP_test.dat
@@ -1,3 +1,0 @@
-YYYY.MM.DD hh:mm:ss,RH (%),T (C),P (Pa)
-2025.10.16 15:27:57,29.60,21.39,98982
-2025.10.16 15:27:58,29.61,21.39,98983
